Ajout AdherentService, AdherentController, hhtpClient POST et GET

+++ src/main/java/fr/moussalli/tpmediatheque/repository/AdherentRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdherentRepository extends JpaRepository<Adherent, Integer> {
+public interface AdherentRepository extends JpaRepository<Adherent, Long> {
 
 }
